Replaced old display system with new Canvas-based display system.
This was done to make it simpler to refresh the visuals the player
sees on screen. Instead of refreshing every game tick, the visuals
refresh automatically whenever a change to any of the active UIs
is detected. While this approach can be costly with lots of small
changes to the UI, there aren't that many UIs in the first place.
Thus, the cost to speed is probably not enough to justify the
complexity necessary for batching changes.